Tweeks for syscall, for future some things needs to be changed, like execve is a bit rough
