+++ arch/x86/kernel/syscalls/sys_exit.c
@@ -1,0 +1,10 @@
+#include <hubble/errno.h>
+#include <hubble/string.h>
+#include <hubble/syscall.h>
+#include <hubble/syscalls.h>
+#include <smp/scheduler.h>
+
+long exit_stub(long a1) {
+  task_exit(a1);
+  __builtin_unreachable();
+}
