+++ arch/x86/kernel/syscalls/sys_read.c
@@ -1,10 +1,3 @@
-/*
- * Syscall: read from a file descriptor.
- *
- * Reads data from the kernel file associated with the given file
- * descriptor into a user-provided buffer.
- */
-
 #include <hubble/syscalls.h>
 
 #include "syscall_entry.h"
@@ -28,9 +21,6 @@ long sys_read(int fd, char *buffer, size_t len) {
 
   if (fd == 0) {
     VFS_File *stdin = vfs_open("/dev/tty0", 0);
-    /* vfs_open() reports failure as ERR_PTR(-errno), not NULL - has to
-     * be checked before handing it to vfs_read(), which dereferences it
-     * unconditionally. */
     if (IS_ERR(stdin) || !stdin)
       return -EIO;
