+++ arch/x86/kernel/syscalls/sys_uid.c
@@ -1,0 +1,8 @@
+#include <hubble/errno.h>
+#include <hubble/printk.h>
+#include <hubble/string.h>
+#include <hubble/syscall.h>
+#include <hubble/syscalls.h>
+
+long sys_getuid(void) { return 0; }
+long sys_geteuid(void) { return 0; }
