+++ arch/x86/kernel/syscalls/syscall.c
@@ -14,13 +14,6 @@ uint64_t syscall_handler_wrapper(registers_t *regs) {
   task->exec.in_syscall_rsp = 0;
   task->exec.in_syscall = true;
 
-  /* fork() needs the raw trap frame (return RIP/RFLAGS, all GPRs) to set up
-   * the child's resume state, which the generic (num, a1..a6) dispatch
-   * below has no room to carry - so it's special-cased here instead of
-   * going through syscall_table. execve() is the same story in reverse:
-   * on success it doesn't return a value at all, it rewrites where the
-   * sysret in syscall_entry.asm lands (RCX / %gs:8), which only the raw
-   * trap frame gives it access to. */
   if (regs->rax == SYS_fork) {
     regs->rax = sys_fork(regs);
   } else if (regs->rax == SYS_execve) {
