+++ init/main.c
@@ -191,12 +191,12 @@ void kmain_thread(void) {
   //   }
 
   //   vfs_close(ext2_file);
-  module_load("/modules/input.ko");
-  module_load("/modules/tty.ko");
-  char *argv[] = {"/busy/busybox", "sh", NULL};
-  task_t *task1 = execv("/busy/busybox", argv, NULL);
-  if (task1 != NULL)
-    scheduler_add_task(task1);
+  //   module_load("/modules/input.ko");
+  //   module_load("/modules/tty.ko");
+  //   char *argv[] = {"/busy/busybox", "sh", NULL};
+  //   task_t *task1 = execv("/busy/busybox", argv, NULL);
+  //   if (task1 != NULL)
+  //     scheduler_add_task(task1);
 
   printk(KERN_INFO "kmain thread done, entering idle loop\n");
